Refactor experiment config actions buttons into separate component
